requirements.txt
h5py==3.16.0
numpy==2.2.6
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
tqdm==4.67.3
wandb==0.25.1